refactor(component): refactor all files in this directory
