+++ src/components/AppBar/index.tsx
@@ -1,0 +1,1 @@
+export * from "./SearchAppBar"
