+++ src/components/Card/Card.tsx
@@ -8,7 +8,7 @@ import Typography from "@mui/material/Typography"
 import { TextField } from "@mui/material"
 
 const imgUrl = "https://brand.zesty.io/zesty-io-logo.svg"
-export default function OutlinedCard({ handleCustomDomain, value, onChange }: any) {
+export const OutlineCard = ({ handleCustomDomain, value, onChange }: any) => {
    const card = (
       <React.Fragment>
          <CardContent>
