+++ src/components/Tabs/index.tsx
@@ -1,0 +1,1 @@
+export * from "./Tabs"
