+++ src/components/Ui/index.tsx
@@ -1,0 +1,2 @@
+export * from "./BackToTop"
+export * from "./StickyHeaderTable"
