+++ src/components/index.tsx
@@ -1,5 +1,8 @@
-export * from "./Headers"
-export * from "./Tabs"
 export * from "./ZestyExplorer"
+export * from "./AppBar"
+export * from "./Card"
+export * from "./Table"
 export * from "./Loader"
-export * from "./SearchAppBar"
+export * from "./Header"
+export * from "./Tabs"
+export * from "./Ui"
