+++ test.html
@@ -5,7 +5,11 @@
       <meta http-equiv="X-UA-Compatible" content="IE=edge" />
       <meta name="viewport" content="width=device-width, initial-scale=1.0" />
       <title>Document</title>
-      <script type="text/javascript" src="./build/bundle-dev.js" defer="defer"></script>
+      <script
+         type="text/javascript"
+         src="/dist/explorer.production.js"
+         defer="defer"
+      ></script>
    </head>
    <body></body>
 </html>
